Added calendar, no data yet... Added the
connection to all the views

+++ AOC-Wk 4/AOC-Wk 3/EventView.h
@@ -20,7 +20,7 @@
 
 @property (nonatomic, strong) UITextField *eventField;
 @property (nonatomic, strong) NSString *eventString;
-@property (nonatomic) int count;
+@property (nonatomic) long count;
 
 
 - (IBAction)hideKey:(id)sender;
